postgresql/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
greenlet
pydantic-settings
PyJWT
